reorganizing, reverting changes to config reader, fixed vault test
