+++ src/functional_tests/.env
@@ -1,0 +1,3 @@
+CG_APPLICATION_NAME=Confluent Gateway Local Test
+CG_ENVIRONMENT=Test
+CG_DB_CONNECTION_STRING=host=localhost user=postgres password=p dbname=db port=5432 sslmode=disable
