+++ src/functional_tests/messaging_test.go
@@ -1,0 +1,1 @@
+package main
