+++ src/functional_tests/mocks/vault_client.go
