+++ src/functional_tests/mocks/vault_client_test.go
